tests/common: ensure all subprocesses are closed
This stops a memleak where an `mfw` subprocess remains active.

Note that an idiosyncrasy in dash as a sh symlink means it is safest to
only close direct subprocesses of a given test script. This is fine,
since `mfw` does not maintain any subprocesses of its own.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=482999